refactor: move actionable banners under proposals
