+++ frontend/src/lib/components/proposals/ActionableProposalsEmpty.svelte
